assignTemp_: Don't assign constant expressions to temporaries
By making sure assignTemp_ doesn't assign constant expressions to a
temporary we work around the weak constant propagation pass in the old
code generator.  Without this change

    thawArray# arr# 0# 16# s#

would not emit a call to MO_Memset with statically know arguments,
resulting in a missed loop unrolling opportunity.

This change is not needed in the new code generator due to its better
constant propagation pass.