codeGen: inline allocation optimization to clone array primops
The inline allocation version is 69% faster than the out-of-line
version, when cloning an array of 16 unit elements on a 64-bit
machine.

Comparing to the old primop implementations isn't straightforward. The
old version had a missing heap check that I discovered during the
development of the new version. Comparing the old and the new version
would requiring fixing the old version, which in turn means
reimplementing the equivalent of MAYBE_CG in StgCmmPrim.

Fixes #8885.